+++ src/GeniusYield/Types/Credential.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE PatternSynonyms #-}
+
 {- |
 Module      : GeniusYield.Types.Credential
 Copyright   : (c) 2023 GYELD GMBH
@@ -6,8 +8,17 @@ Maintainer  : support@geniusyield.co
 Stability   : develop
 -}
 module GeniusYield.Types.Credential (
+  -- * Credential.
+  GYCredential (..),
+  credentialToHexText,
+  credentialToLedger,
+  credentialFromLedger,
+  credentialToPlutus,
+
   -- * Payment credential.
-  GYPaymentCredential (..),
+  GYPaymentCredential,
+  pattern GYPaymentCredentialByKey,
+  pattern GYPaymentCredentialByScript,
   paymentCredentialToApi,
   paymentCredentialFromApi,
   paymentCredentialToLedger,
@@ -17,7 +28,9 @@ module GeniusYield.Types.Credential (
   paymentCredentialToBech32,
 
   -- * Stake credential.
-  GYStakeCredential (..),
+  GYStakeCredential,
+  pattern GYStakeCredentialByKey,
+  pattern GYStakeCredentialByScript,
   stakeCredentialToApi,
   stakeCredentialFromApi,
   stakeCredentialToLedger,
@@ -31,16 +44,14 @@ import Cardano.Api.Ledger qualified as Ledger
 import Cardano.Api.Shelley qualified as Api
 import Data.Hashable (Hashable (..))
 import Data.Text (Text)
-import GeniusYield.Imports ((>>>))
+import Data.Text qualified as Text
+import GeniusYield.Types.KeyHash
+import GeniusYield.Types.KeyRole
 import GeniusYield.Types.PaymentKeyHash (
-  GYPaymentKeyHash,
   paymentKeyHashFromApi,
-  paymentKeyHashFromLedger,
   paymentKeyHashToApi,
-  paymentKeyHashToLedger,
-  paymentKeyHashToPlutus,
  )
-import GeniusYield.Types.PubKeyHash (AsPubKeyHash (fromPubKeyHash, toPubKeyHash))
+import GeniusYield.Types.PubKeyHash (AsPubKeyHash (toPubKeyHash), pubKeyHashToPlutus)
 import GeniusYield.Types.Script (
   GYScriptHash,
   scriptHashFromApi,
@@ -50,7 +61,6 @@ import GeniusYield.Types.Script (
   scriptHashToPlutus,
   stakeValidatorHashFromApi,
   stakeValidatorHashToApi,
-  stakeValidatorHashToPlutus,
  )
 import GeniusYield.Types.StakeKeyHash (
   GYStakeKeyHash,
@@ -61,18 +71,28 @@ import GeniusYield.Utils (serialiseToBech32WithPrefix)
 import PlutusLedgerApi.V1 qualified as Plutus (Credential (..))
 import Text.Printf qualified as Printf
 
--- | Payment credential.
-data GYPaymentCredential
-  = GYPaymentCredentialByKey !GYPaymentKeyHash
-  | GYPaymentCredentialByScript !GYScriptHash
-  deriving (Show, Eq, Ord)
+{- $setup
 
-instance Printf.PrintfArg GYPaymentCredential where
-  formatArg (GYPaymentCredentialByKey pkh) = Printf.formatArg $ "Payment key credential: " <> Api.serialiseToRawBytesHexText (paymentKeyHashToApi pkh)
-  formatArg (GYPaymentCredentialByScript sh) = Printf.formatArg $ "Payment script credential: " <> Api.serialiseToRawBytesHexText (scriptHashToApi sh)
+>>> :set -XOverloadedStrings -XTypeApplications -XDataKinds
+>>> import qualified Text.Printf                as Printf
+>>> import GeniusYield.Types.KeyRole
+>>> import GeniusYield.Types.KeyHash
+>>> let pkh :: GYKeyHash 'GYKeyRolePayment = "ec91ac77b581ba928db86cd91d11e64032450677c6b80748ce0b9a81"
+>>> let pcred = GYCredentialByKey pkh
+>>> let scred = GYCredentialByScript "464eeee89f05aff787d40045af2a40a83fd96c513197d32fbc54ff02"
+-}
 
-instance Hashable GYPaymentCredential where
-  hashWithSalt salt cred = hashWithSalt salt $ paymentCredentialToHexText cred
+{- | Payment credential.
+@type GYPaymentCredential = GYCredential 'GYKeyRolePayment@
+-}
+type GYPaymentCredential = GYCredential 'GYKeyRolePayment
+
+pattern GYPaymentCredentialByKey :: GYKeyHash 'GYKeyRolePayment -> GYPaymentCredential
+pattern GYPaymentCredentialByKey kh = GYCredentialByKey kh
+pattern GYPaymentCredentialByScript :: GYScriptHash -> GYPaymentCredential
+pattern GYPaymentCredentialByScript sh = GYCredentialByScript sh
+
+{-# COMPLETE GYPaymentCredentialByKey, GYPaymentCredentialByScript #-}
 
 -- | Convert @GY@ type to corresponding type in @cardano-node@ library.
 paymentCredentialToApi :: GYPaymentCredential -> Api.PaymentCredential
@@ -86,41 +106,35 @@ paymentCredentialFromApi (Api.PaymentCredentialByScript sh) = GYPaymentCredentia
 
 -- | Convert to corresponding ledger representation.
 paymentCredentialToLedger :: GYPaymentCredential -> Ledger.Credential Ledger.Payment Ledger.StandardCrypto
-paymentCredentialToLedger pc = case pc of
-  GYPaymentCredentialByKey kh -> Ledger.KeyHashObj $ paymentKeyHashToLedger kh
-  GYPaymentCredentialByScript sh -> Ledger.ScriptHashObj $ scriptHashToLedger sh
+paymentCredentialToLedger = credentialToLedger
 
 paymentCredentialFromLedger :: Ledger.Credential Ledger.Payment Ledger.StandardCrypto -> GYPaymentCredential
-paymentCredentialFromLedger c = case c of
-  Ledger.KeyHashObj kh -> GYPaymentCredentialByKey $ paymentKeyHashFromLedger kh
-  Ledger.ScriptHashObj sh -> GYPaymentCredentialByScript $ scriptHashFromLedger sh
+paymentCredentialFromLedger = credentialFromLedger
 
 -- | Convert @GY@ type to corresponding type in @plutus@ library.
 paymentCredentialToPlutus :: GYPaymentCredential -> Plutus.Credential
-paymentCredentialToPlutus (GYPaymentCredentialByKey pkh) = Plutus.PubKeyCredential (paymentKeyHashToPlutus pkh)
-paymentCredentialToPlutus (GYPaymentCredentialByScript sh) = Plutus.ScriptCredential (scriptHashToPlutus sh)
+paymentCredentialToPlutus = credentialToPlutus
 
 -- | Get hexadecimal value of payment credential.
 paymentCredentialToHexText :: GYPaymentCredential -> Text
-paymentCredentialToHexText =
-  \case
-    GYPaymentCredentialByKey pkh -> Api.serialiseToRawBytesHexText (paymentKeyHashToApi pkh)
-    GYPaymentCredentialByScript sh -> Api.serialiseToRawBytesHexText (scriptHashToApi sh)
+paymentCredentialToHexText = credentialToHexText
 
 -- | Get the bech32 encoding for the given credential.
 paymentCredentialToBech32 :: GYPaymentCredential -> Text
 paymentCredentialToBech32 (GYPaymentCredentialByKey pkh) = serialiseToBech32WithPrefix "addr_vkh" $ paymentKeyHashToApi pkh
 paymentCredentialToBech32 (GYPaymentCredentialByScript sh) = serialiseToBech32WithPrefix "addr_shared_vkh" $ scriptHashToApi sh
 
--- | Stake credential.
-data GYStakeCredential
-  = GYStakeCredentialByKey !GYStakeKeyHash
-  | GYStakeCredentialByScript !GYScriptHash
-  deriving (Show, Eq, Ord)
+{- | Stake credential.
+@type GYStakeCredential = GYCredential 'GYKeyRoleStaking@
+-}
+type GYStakeCredential = GYCredential 'GYKeyRoleStaking
 
-instance Printf.PrintfArg GYStakeCredential where
-  formatArg (GYStakeCredentialByKey skh) = Printf.formatArg $ "Stake key credential: " <> Api.serialiseToRawBytesHexText (stakeKeyHashToApi skh)
-  formatArg (GYStakeCredentialByScript sh) = Printf.formatArg $ "Stake script credential: " <> Api.serialiseToRawBytesHexText (stakeValidatorHashToApi sh)
+pattern GYStakeCredentialByKey :: GYStakeKeyHash -> GYStakeCredential
+pattern GYStakeCredentialByKey kh = GYCredentialByKey kh
+pattern GYStakeCredentialByScript :: GYScriptHash -> GYStakeCredential
+pattern GYStakeCredentialByScript sh = GYCredentialByScript sh
+
+{-# COMPLETE GYStakeCredentialByKey, GYStakeCredentialByScript #-}
 
 -- | Convert @GY@ type to corresponding type in @cardano-api@ library.
 stakeCredentialToApi :: GYStakeCredential -> Api.StakeCredential
@@ -134,20 +148,59 @@ stakeCredentialFromApi (Api.StakeCredentialByScript sh) = GYStakeCredentialByScr
 
 -- | Convert to corresponding ledger type.
 stakeCredentialToLedger :: GYStakeCredential -> Ledger.Credential Ledger.Staking Ledger.StandardCrypto
-stakeCredentialToLedger = stakeCredentialToApi >>> Api.toShelleyStakeCredential
+stakeCredentialToLedger = credentialToLedger
 
 -- | Convert from corresponding ledger type.
 stakeCredentialFromLedger :: Ledger.Credential Ledger.Staking Ledger.StandardCrypto -> GYStakeCredential
-stakeCredentialFromLedger = Api.fromShelleyStakeCredential >>> stakeCredentialFromApi
+stakeCredentialFromLedger = credentialFromLedger
 
 -- | Convert @GY@ type to corresponding type in @plutus@ library.
 stakeCredentialToPlutus :: GYStakeCredential -> Plutus.Credential
-stakeCredentialToPlutus (GYStakeCredentialByKey pkh) = Plutus.PubKeyCredential (paymentKeyHashToPlutus $ fromPubKeyHash $ toPubKeyHash pkh)
-stakeCredentialToPlutus (GYStakeCredentialByScript sh) = Plutus.ScriptCredential (stakeValidatorHashToPlutus sh)
+stakeCredentialToPlutus = credentialToPlutus
 
 -- | Get hexadecimal value of stake credential.
 stakeCredentialToHexText :: GYStakeCredential -> Text
-stakeCredentialToHexText =
+stakeCredentialToHexText = credentialToHexText
+
+-- | Credential.
+data GYCredential (kr :: GYKeyRole)
+  = GYCredentialByKey !(GYKeyHash kr)
+  | GYCredentialByScript !GYScriptHash
+  deriving stock (Show, Eq, Ord)
+
+{- | Get hexadecimal value of credential.
+
+>>> credentialToHexText pcred
+"ec91ac77b581ba928db86cd91d11e64032450677c6b80748ce0b9a81"
+
+>>> credentialToHexText scred
+"464eeee89f05aff787d40045af2a40a83fd96c513197d32fbc54ff02"
+-}
+credentialToHexText :: GYCredential kr -> Text
+credentialToHexText =
   \case
-    GYStakeCredentialByKey skh -> Api.serialiseToRawBytesHexText (stakeKeyHashToApi skh)
-    GYStakeCredentialByScript sh -> Api.serialiseToRawBytesHexText (stakeValidatorHashToApi sh)
+    GYCredentialByKey kh -> keyHashToRawBytesHexText kh
+    GYCredentialByScript sh -> Api.serialiseToRawBytesHexText (scriptHashToApi sh)
+
+{- |
+>>> Printf.printf "%s\n" $ pcred
+Key credential (GYKeyRolePayment): ec91ac77b581ba928db86cd91d11e64032450677c6b80748ce0b9a81
+-}
+instance SingGYKeyRoleI kr => Printf.PrintfArg (GYCredential kr) where
+  formatArg (GYCredentialByKey kh) = Printf.formatArg $ "Key credential (" <> show (fromSingGYKeyRole $ singGYKeyRole @kr) <> "): " <> Text.unpack (keyHashToRawBytesHexText kh)
+  formatArg (GYCredentialByScript sh) = Printf.formatArg $ "Script credential: " <> Api.serialiseToRawBytesHexText (scriptHashToApi sh)
+
+instance Hashable (GYCredential kr) where
+  hashWithSalt salt cred = hashWithSalt salt $ credentialToHexText cred
+
+credentialToLedger :: GYCredential kr -> Ledger.Credential (GYKeyRoleToLedger kr) Ledger.StandardCrypto
+credentialToLedger (GYCredentialByKey kh) = Ledger.KeyHashObj $ keyHashToLedger kh
+credentialToLedger (GYCredentialByScript sh) = Ledger.ScriptHashObj $ scriptHashToLedger sh
+
+credentialFromLedger :: Ledger.Credential (GYKeyRoleToLedger kr) Ledger.StandardCrypto -> GYCredential kr
+credentialFromLedger (Ledger.KeyHashObj kh) = GYCredentialByKey $ keyHashFromLedger kh
+credentialFromLedger (Ledger.ScriptHashObj sh) = GYCredentialByScript $ scriptHashFromLedger sh
+
+credentialToPlutus :: GYCredential kr -> Plutus.Credential
+credentialToPlutus (GYCredentialByKey kh) = Plutus.PubKeyCredential (pubKeyHashToPlutus $ toPubKeyHash kh)
+credentialToPlutus (GYCredentialByScript sh) = Plutus.ScriptCredential (scriptHashToPlutus sh)